fix: changed deltaTime type to double, removed unnecessary typecasting

+++ src/main.cpp
@@ -54,8 +54,8 @@ int main(int argc, char* argv[]) {
 
     Uint64 now = SDL_GetPerformanceCounter();
     Uint64 last = 0;
-    float dt = 0.0f;
-    
+    double dt = 0.0;
+
     while (running) {
         last = now;
         now = SDL_GetPerformanceCounter();
@@ -63,7 +63,7 @@ int main(int argc, char* argv[]) {
 
         // GAME SYSTEM (EVENTS, UPDATION, RENDERING)
         running = game.handleEvents();
-        game.tick(static_cast<float>(dt));
+        game.tick(static_cast<double>(dt));
         render(renderer, game);
     }
